ok pour la modération

+++ backend/src/resolvers/report.resolver.ts
@@ -7,11 +7,12 @@ import { MyContext } from "..";
 import UserEntity from "../entities/user.entity";
 import { ReportStatus } from "../enum/reportStatus.enum";
 import { ReportEntity } from "../entities/reportEntity.entity";
+import JoinUserParkourNoteEntity from "../entities/joinUserParkourNote.entity";
 
 @Resolver()
 export default class ReportResolver {
   // page user by admin
-  // @Authorized("ADMIN")
+  @Authorized("ADMIN")
   @Query(() => UserEntity)
   async getUserByIdForPageReport(@Arg("userId") userId: string) {
     const result = await new ReportService().getUserByIdForPageReport(userId);
@@ -20,7 +21,7 @@ export default class ReportResolver {
   }
 
   // page search reports
-  // @Authorized("ADMIN")
+  @Authorized("ADMIN")
   @Query(() => [ReportEntity])
   async getReportsBySearch(@Arg("status") status: ReportStatus) {
     const result: ReportEntity[] = await new ReportService().getReportsBySearch(
@@ -32,24 +33,32 @@ export default class ReportResolver {
   // ---
 
   // user sur malfrat comm
-  // @Authorized("CLIENT")
+  @Authorized("CLIENT")
   @Mutation(() => MessageEntity)
   async reportNote(
-    @Arg("malfrat_id") malfrat_id: string,
-    @Arg("parkour_id") parkour_id: number,
+    @Arg("malfratId") malfratId: string,
+    @Arg("parkourId") parkourId: number,
     @Ctx() ctx: MyContext,
     @Arg("commentaire") commentaire: string
   ) {
     const returnMessage = new MessageEntity();
 
     if (ctx.user) {
-      // cré le report
-      await new ReportService().reportNoteByUserIdAndParkourId(
-        malfrat_id,
-        parkour_id,
-        ctx.user.id,
+      const reportExist = await new ReportService().isReportExist(
+        malfratId,
+        parkourId,
         commentaire
       );
+
+      if (!reportExist) {
+        // cré le report
+        await new ReportService().reportNoteByUserIdAndParkourId(
+          malfratId,
+          parkourId,
+          ctx.user.id,
+          commentaire
+        );
+      }
 
       // ajoute 1 au nb de report ajoutés par le user
       await new ReportService().addOneNbReportAjouteByToken(ctx.user);
@@ -62,20 +71,16 @@ export default class ReportResolver {
     return returnMessage;
   }
 
+  // ---
+
   // admin sur malfrat comm - laisse
-  // @Authorized("ADMIN")
+  @Authorized("ADMIN")
   @Mutation(() => MessageEntity)
-  async letNoteFromReport(
-    @Arg("user_id") malfrat_id: string,
-    @Arg("parkour_id") parkour_id: number
-  ) {
+  async letNote(@Arg("reportId") reportId: number) {
     const returnMessage = new MessageEntity();
 
-    // on laisse le comm => on enlève les reports (c'est tout)
-    await new ReportService().deleteReportsByUserIdAndParkourId(
-      malfrat_id,
-      parkour_id
-    );
+    // on laisse le comm => on enlève le report (c'est tout)
+    await new ReportService().deleteReportByReportId(reportId);
 
     returnMessage.message = "Vous venez de laisser le message d'un utilisateur";
     returnMessage.success = true;
@@ -83,68 +88,38 @@ export default class ReportResolver {
     return returnMessage;
   }
 
-  // admin sur malfrat comm - laisse (car modifié ou supprimé) mais ajoute un reportValide sur le user
-  // @Authorized("ADMIN")
-  @Mutation(() => MessageEntity)
-  async addOneReportValideFromReport(
-    @Arg("user_id") malfrat_id: string,
-    @Arg("parkour_id") parkour_id: number,
-    @Arg("commentaire") commentaire: string
-  ) {
-    const returnMessage = new MessageEntity();
-
-    // ajoute 1 nbReportValide à ce user
-    await new ReportService().addOneReportValideForUser(malfrat_id);
-
-    // pour éviter les doublons
-    await new ReportService().keepOnlyOneReport(
-      malfrat_id,
-      parkour_id,
-      commentaire
-    );
-    await new ReportService().modifyStatusReport(
-      malfrat_id,
-      parkour_id,
-      ReportStatus.VU_ET_LAISSE_MODIF
-    );
-
-    returnMessage.message =
-      "Vous venez de laisser le message d'un utilisateur mais ajouté un report sur son profil";
-    returnMessage.success = true;
-
-    return returnMessage;
-  }
-
   // admin sur malfrat comm - delete
-  // @Authorized("ADMIN")
+  @Authorized("ADMIN")
   @Mutation(() => MessageEntity)
-  async deleteNoteFromReport(
-    @Arg("user_id") malfrat_id: string,
-    @Arg("parkour_id") parkour_id: number,
+  async deleteNoteAndAddOneReportValide(
+    @Arg("malfratId") malfratId: string,
+    @Arg("parkourId") parkourId: number,
+    @Arg("reportId") reportId: number,
     @Arg("commentaire") commentaire: string
   ) {
     const returnMessage = new MessageEntity();
 
-    // supprime la note
-    await new JoinUserParkourNoteService().deleteNoteByUserIdAndParkourId(
-      malfrat_id,
-      parkour_id
+    const joinUserParkourNote: JoinUserParkourNoteEntity | null =
+      await new JoinUserParkourNoteService().getNoteByUserIdAndParkourIdOrNull(
+        malfratId,
+        parkourId
+      );
+
+    if (joinUserParkourNote && joinUserParkourNote.commentaire == commentaire) {
+      // supprime la note
+      await new JoinUserParkourNoteService().deleteNoteByUserIdAndParkourId(
+        malfratId,
+        parkourId
+      );
+    }
+
+    await new ReportService().modifyStatusReport(
+      reportId,
+      ReportStatus.SUPPRIME
     );
 
     // ajoute 1 nbReportValide à ce user
-    await new ReportService().addOneReportValideForUser(malfrat_id);
-
-    // pour éviter les doublons (ne marche pas de mettre un flag sur le joinNote)
-    await new ReportService().keepOnlyOneReport(
-      malfrat_id,
-      parkour_id,
-      commentaire
-    );
-    await new ReportService().modifyStatusReport(
-      malfrat_id,
-      parkour_id,
-      ReportStatus.VU_ET_SUPPRIME
-    );
+    await new ReportService().addOneReportValideForUser(malfratId);
 
     returnMessage.message =
       "Vous venez de supprimer le message d'un utilisateur";
@@ -153,15 +128,50 @@ export default class ReportResolver {
     return returnMessage;
   }
 
+  // admin sur malfrat comm - delete sans passer par report
+  @Authorized("ADMIN")
+  @Mutation(() => MessageEntity)
+  async deleteNoteAndAddOneReportValideAndCreateReport(
+    @Arg("malfratId") malfratId: string,
+    @Arg("parkourId") parkourId: number,
+    @Ctx() ctx: MyContext,
+    @Arg("commentaire") commentaire: string
+  ) {
+    const returnMessage = new MessageEntity();
+
+    if (ctx.user) {
+      await new JoinUserParkourNoteService().deleteNoteByUserIdAndParkourId(
+        malfratId,
+        parkourId
+      );
+
+      // cré le repport
+      await new ReportService().createDeleteReport(
+        malfratId,
+        parkourId,
+        ctx.user.id,
+        commentaire
+      );
+
+      // ajoute 1 nbReportValide à ce user
+      await new ReportService().addOneReportValideForUser(malfratId);
+
+      returnMessage.message =
+        "Vous venez de supprimer le message d'un utilisateur";
+      returnMessage.success = true;
+    }
+
+    return returnMessage;
+  }
+
   // admin sur malfrat compte
   // @Authorized("ADMIN")
   @Mutation(() => MessageEntity)
-  async deleteUserByReport(@Arg("user_id") malfrat_id: string) {
+  async deleteUserByAdmin(@Arg("malfratId") malfratId: string) {
     const returnMessage = new MessageEntity();
 
-    const user = await new UserService().getUserById(malfrat_id);
+    const user = await new UserService().getUserById(malfratId);
     // (possiblement garder son email pour ne pas le revoir)
-    await new ReportService().supprimeAllReportsByUserId(malfrat_id);
     await new UserService().deleteUser(user);
 
     returnMessage.message = "Vous venez de supprimer un utilisateur";
